# Slightly faster IQ4_XS_R4 on AVX2